requirements.txt
flask
gunicorn
sentence-transformers
faiss-cpu
google-generativeai
python-dotenv
